尝试解决debug通过pip无法正常使用问题

+++ dingraia/debug/__main__.py
@@ -1,33 +1,49 @@
 import signal
-import time
+import platform
 import os
 import sys
 import argparse
 import subprocess
+import time
+
 from watchdog.observers import Observer
 from watchdog.events import FileSystemEventHandler
 from ..log import logger
 
 python_executable = sys.executable
 isOnReload = False
+logger = logger.switch_logger(2)
+waitAfterLastReload = 5
+waitAfterShutdown = 1
+lastReloadTime = time.time()
 
 
-class MyHandler(FileSystemEventHandler):
+class FileChangedHandler(FileSystemEventHandler):
 
     def on_modified(self, event):
+        global lastReloadTime
         if not isOnReload:
             if list(filter(lambda x: event.src_path.endswith(x), watch_file_type)):
                 logger.warning(f'文件 {event.src_path} 发生了变动')
-                your_function()
+                if time.time() - lastReloadTime < waitAfterLastReload:
+                    logger.warning(f'距离上次重载时间过短, 设定值为 {waitAfterLastReload} 秒')
+                    return
+                lastReloadTime = time.time()
+                stop_program()
 
 
-def your_function():
+def stop_program():
     global exit_code
     global isOnReload
+    if isOnReload:
+        return
     isOnReload = True
     exit_code = -114514
     logger.info("正在重载...")
-    process.send_signal(signal.CTRL_C_EVENT)
+    if platform.system() == "Windows":
+        process.send_signal(signal.CTRL_C_EVENT)
+    else:
+        process.send_signal(signal.SIGINT)
 
 
 if __name__ == "__main__":
@@ -45,7 +61,7 @@ if __name__ == "__main__":
         watch_file_type = []
     else:
         watch_file_type = [watch_file_type]
-    event_handler = MyHandler()
+    event_handler = FileChangedHandler()
     observer = Observer()
     observer.schedule(event_handler, path, recursive=True)
     observer.start()
@@ -53,14 +69,19 @@ if __name__ == "__main__":
     exit_code = None
     try:
         while exit_code in (None, -114514):
-            process = subprocess.Popen([python_executable, python_file] + python_args, shell=True)
-            logger.info(f"正在运行 {python_file}, PID {process.pid}")
+            cmd = [python_executable, python_file] + python_args
+            process = subprocess.Popen(cmd, shell=False)
+            logger.info(f"启动命令: {cmd}")
+            logger.info(f"正在运行 {python_file}, Python PID {process.pid}")
             isOnReload = False
             try:
                 process.wait()
             except KeyboardInterrupt:
                 logger.info("正在退出, 等待程序结束...")
-                process.send_signal(signal.CTRL_C_EVENT)
+                if platform.system() == "Windows":
+                    process.send_signal(signal.CTRL_C_EVENT)
+                else:
+                    process.send_signal(signal.SIGINT)
                 process.wait()
                 exit_code = 1
                 logger.info("程序已退出")
@@ -75,6 +96,7 @@ if __name__ == "__main__":
                     exit_code = process.returncode
                 else:
                     exit_code = None
+                    time.sleep(waitAfterShutdown)
     except KeyboardInterrupt:
         observer.stop()
     observer.join()
